+++ sdk/storage/azure-storage-file-share/src/test/java/com/azure/storage/file/share/APISpec.groovy
@@ -39,7 +39,7 @@ class APISpec extends Specification {
     String connectionString
 
     // If debugging is enabled, recordings cannot run as there can only be one proxy at a time.
-    static boolean enableDebugging = false
+    static boolean enableDebugging = true
 
     /**
      * Setup the File service clients commonly used for the API tests.